move logging stuff to ledger/error
