+++ ledger/ledger-api-common/src/test/suite/scala/com/digitalasset/ledger/api/validation/ErrorFactoriesSpec.scala
@@ -11,16 +11,9 @@ import ch.qos.logback.classic.Level
 import com.daml.error.definitions.LedgerApiErrors.RequestValidation.InvalidDeduplicationPeriodField.ValidMaxDeduplicationFieldKey
 import com.daml.error.definitions.{CommonErrors, IndexErrors, LedgerApiErrors}
 import com.daml.error.utils.ErrorDetails
-import com.daml.error.{
-  ContextualizedErrorLogger,
-  DamlContextualizedErrorLogger,
-  DamlError,
-  ErrorAssertionsWithLogCollectorAssertions,
-  ErrorCode,
-}
+import com.daml.error.{ContextualizedErrorLogger, DamlContextualizedErrorLogger, DamlError, ErrorAssertionsWithLogCollectorAssertions, ErrorCode, LogCollector, LogCollectorAssertions}
 import com.daml.lf.data.Ref
-import com.daml.platform.testing.LogCollector.ExpectedLogEntry
-import com.daml.platform.testing.{LogCollector, LogCollectorAssertions}
+import com.daml.error.LogCollector.ExpectedLogEntry
 import com.google.rpc._
 import io.grpc.Status.Code
 import io.grpc.StatusRuntimeException
