+++ test-common/BUILD.bazel
@@ -61,10 +61,6 @@ alias(
             name = "test-common-%s" % target,
             srcs = glob(["src/main/scala/**/*.scala"]),
             scala_deps = [
-                "@maven//:org_scalatest_scalatest_shouldmatchers",
-                "@maven//:org_scalatest_scalatest_core",
-                "@maven//:org_scalactic_scalactic",
-                "@maven//:org_scalatest_scalatest_matchers_core",
             ],
             tags = ["maven_coordinates=com.daml:test-common-%s:__VERSION__" % target],
             visibility = [
@@ -72,12 +68,7 @@ alias(
             ],
             deps = [
                 "//ledger-api/grpc-definitions:ledger_api_proto_scala",
-                "//libs-scala/scala-utils",
                 "//libs-scala/timer-utils",
-                "@maven//:ch_qos_logback_logback_classic",
-                "@maven//:ch_qos_logback_logback_core",
-                "@maven//:org_scalatest_scalatest_compatible",
-                "@maven//:org_slf4j_slf4j_api",
             ],
         ),
     ]
